+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
+- feature/1200523060528388 - Add recursive flag - Add CallableCollection parameters on LocalFileFinderAdapter 
 - feature/1200491234707795 - Adapt new libraries
 - feature/1200366404543322 - Modification which Integrates validator reset capability / Max depth directory validator